+++ src/core/settings_manager.cpp
@@ -130,6 +130,12 @@ void SettingsManager::setShowHints(bool value) {
     notifyIfChanged(old);
 }
 
+void SettingsManager::setShowSessionTimer(bool value) {
+    auto old = settings_;
+    settings_.show_session_timer = value;
+    notifyIfChanged(old);
+}
+
 void SettingsManager::setCollectDetailedStats(bool value) {
     auto old = settings_;
     settings_.collect_detailed_stats = value;
@@ -202,6 +208,9 @@ void SettingsManager::load() {
             if (auto v = display["show_hints"]) {
                 settings_.show_hints = v.as<bool>();
             }
+            if (auto v = display["show_session_timer"]) {
+                settings_.show_session_timer = v.as<bool>();
+            }
             if (auto v = display["collect_detailed_stats"]) {
                 settings_.collect_detailed_stats = v.as<bool>();
             }
@@ -245,6 +254,7 @@ void SettingsManager::save() const {
         YAML::Node display;
         display["show_conflicts"] = settings_.show_conflicts;
         display["show_hints"] = settings_.show_hints;
+        display["show_session_timer"] = settings_.show_session_timer;
         display["collect_detailed_stats"] = settings_.collect_detailed_stats;
         display["encrypt_detailed_stats"] = settings_.encrypt_detailed_stats;
         root["display"] = display;